requirements.txt
async-timeout==5.0.1
certifi==2025.7.14
charset-normalizer==3.4.2
idna==3.10
pyTelegramBotAPI==4.27.0
python-dotenv==1.1.1
redis==6.2.0
requests==2.32.4
six==1.17.0
sseclient==0.0.27
urllib3==2.5.0
